Rename dfns.json, separate tests that need an API key
Tests that require a W3C API key cannot run on pull requests from forked
repositories. This update splits them to a separate file, and create two
separate workflows:
- one for push actions that runs all tests
- one for pull requests that runs all tests that don't require a W3C API key

dfns.json renamed to definitions.json, and top-level key in that file renamed
to `proptype`.
