+++ .github/workflows/lint-pr.yml
@@ -1,4 +1,4 @@
-on: push
+on: pull_request
 name: lint
 jobs:
   lint:
@@ -8,7 +8,6 @@ jobs:
     - uses: actions/setup-node@v1
       with:
         node-version: 12.x
-    - run: echo "${{ secrets.CONFIG_JSON }}" | base64 --decode > config.json
     - run: npm ci
-    - run: npm run test
+    - run: npm run test-pr
     - run: npm run lint
